Updated sample GuiApp.PrismAvalonia
