+++ sandbox/GuiApp.PrismAvalonia.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31605.320
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32804.467
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GuiApp-Avalonia", "GuiApp-Avalonia\GuiApp-Avalonia.csproj", "{98F164D1-51CF-4BC1-8F9D-8DFB541E57F5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GuiApp.PrismAvalonia", "GuiApp.PrismAvalonia\GuiApp.PrismAvalonia.csproj", "{98F164D1-51CF-4BC1-8F9D-8DFB541E57F5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
