+++ sandbox/GuiApp.PrismAvalonia/Assets/avalonia-logo.ico
