+++ sandbox/GuiApp.PrismAvalonia/GuiApp.PrismAvalonia.csproj
