+++ sandbox/GuiApp.PrismAvalonia/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "Learn.PrismAvalonia": {
+    "GuiApp.PrismAvalonia": {
       "commandName": "Project"
     },
     "WSL": {
