+++ sandbox/GuiApp.PrismAvalonia/ViewModels/DashboardViewModel.cs
@@ -1,0 +1,10 @@
+﻿namespace GuiApp.PrismAvalonia.ViewModels
+{
+  public class DashboardViewModel : ViewModelBase
+  {
+    public DashboardViewModel()
+    {
+      Title = "Dashboard View!";
+    }
+  }
+}
