+++ sandbox/GuiApp.PrismAvalonia/ViewModels/SettingsViewModel.cs
@@ -1,0 +1,10 @@
+﻿namespace GuiApp.PrismAvalonia.ViewModels
+{
+  public class SettingsViewModel : ViewModelBase
+  {
+    public SettingsViewModel()
+    {
+      Title = "Settings View!";
+    }
+  }
+}
