+++ sandbox/GuiApp.PrismAvalonia/Views/DashboardView.axaml.cs
@@ -1,8 +1,8 @@
-using Avalonia;
+﻿using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace Learn.PrismAvalonia.Views
+namespace GuiApp.PrismAvalonia.Views
 {
   public partial class DashboardView : UserControl
   {
