+++ sandbox/GuiApp.PrismAvalonia/Views/SettingsView.axaml
@@ -5,6 +5,9 @@
              xmlns:prism="http://prismlibrary.com/"
              prism:ViewModelLocator.AutoWireViewModel="True"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="Learn.PrismAvalonia.Views.SettingsView">
-  Welcome to Avalonia!
+             x:Class="GuiApp.PrismAvalonia.Views.SettingsView">
+  <StackPanel VerticalAlignment="Center" HorizontalAlignment="Center">>
+    <Label>Welcome to Avalonia!</Label>
+    <TextBlock Text="{Binding Title}" />
+  </StackPanel>  
 </UserControl>
