+++ sandbox/GuiApp.PrismAvalonia/Views/ShellWindow.axaml.cs
@@ -2,7 +2,7 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace Learn.PrismAvalonia.Views
+namespace GuiApp.PrismAvalonia.Views
 {
   public partial class ShellWindow : Window
   {
